+++ src/java/org/malbino/orion/servlets/HistorialEconomico.java
@@ -19,6 +19,7 @@ import com.itextpdf.text.pdf.PdfPCell;
 import com.itextpdf.text.pdf.PdfPTable;
 import com.itextpdf.text.pdf.PdfWriter;
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -234,7 +235,7 @@ public class HistorialEconomico extends HttpServlet {
             cell.setColspan(10);
             table.addCell(cell);
             
-            cell = new PdfPCell(new Phrase(Redondeo.formatear_csm(detalle.getCantidad()), NORMAL));
+            cell = new PdfPCell(new Phrase(Redondeo.formatear_csm(BigDecimal.valueOf(detalle.getCantidad())), NORMAL));
             cell.setHorizontalAlignment(PdfPCell.ALIGN_RIGHT);
             cell.setColspan(10);
             table.addCell(cell);
@@ -249,7 +250,7 @@ public class HistorialEconomico extends HttpServlet {
             cell.setColspan(25);
             table.addCell(cell);
 
-            cell = new PdfPCell(new Phrase(Redondeo.formatear_csm(detalle.getSubtotal()), NORMAL));
+            cell = new PdfPCell(new Phrase(Redondeo.formatear_csm(BigDecimal.valueOf(detalle.getSubtotal())), NORMAL));
             cell.setHorizontalAlignment(PdfPCell.ALIGN_RIGHT);
             cell.setColspan(10);
             table.addCell(cell);
